# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/commerce/orders/{id}/refunds": {
      "post": {
        "operationId": "CommerceController_requestRefund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "Commerce"
        ]
      }
    },
...
    "/api/v1/checkout": {
      "post": {
        "operationId": "CheckoutController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ticketTypeId",
                  "quantity",
                  "phoneNumber"
                ],
                "properties": {
                  "ticketTypeId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "254712345678"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Checkout initiated, STK Push sent.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "paymentId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "checkoutSessionId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "example": "PENDING"
                    },
                    "amount": {
                      "type": "string",
                      "example": "3000.0000"
                    },
                    "currency": {
                      "type": "string",
                      "example": "KES"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (e.g., invalid phone format)."
          },
          "401": {
            "description": "Missing or invalid authentication token."
          },
          "409": {
            "description": "Insufficient ticket inventory."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initiate a ticket purchase and M-Pesa STK Push",
        "tags": [
          "Commerce - Checkout"
        ]
      }
    },
    "/api/v1/media/search": {
      "get": {
        "operationId": "MediaController_search",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ANIME",
                "MANGA"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching media.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Search for anime or manga across the unified media layer",
        "tags": [
          "Content - Media"
        ]
      }
    },
    "/api/v1/media/trending": {
      "get": {
        "operationId": "MediaController_getTrending",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ANIME",
                "MANGA"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trending media feed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get currently trending anime/manga",
        "tags": [
          "Content - Media"
        ]
      }
    },
    "/api/v1/media/airing": {
      "get": {
        "operationId": "MediaController_getAiring",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ANIME",
                "MANGA"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Currently airing media feed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get currently releasing anime",
        "tags": [
          "Content - Media"
        ]
      }
    },
    "/api/v1/media/seasonal": {
      "get": {
        "operationId": "MediaController_getSeasonal",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ANIME",
                "MANGA"
              ],
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "season",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "WINTER",
                "SPRING",
                "SUMMER",
                "FALL"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Seasonal media feed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get anime/manga by specific season and year",
        "tags": [
          "Content - Media"
        ]
      }
    },
    "/api/v1/media/{provider}/{externalId}": {
      "get": {
        "operationId": "MediaController_getById",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "example": "ANILIST",
              "type": "string"
            }
          },
          {
            "name": "externalId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "154587",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ANIME",
                "MANGA"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Specific media item.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleMediaResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Media not found."
          }
        },
        "summary": "Retrieve specific media by provider identity",
        "tags": [
          "Content - Media"
        ]
      }
    }
  },
  "info": {
    "title": "Animanga Platform API",
    "description": "Strictly defined contracts for Commerce, Events, and Identity",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "ScannerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-scanner-api-key"
      }
    },
    "schemas": {
      "MediaTitleDto": {
        "type": "object",
        "properties": {
          "romaji": {
            "type": "string"
          },
          "english": {
            "type": "string"
          },
          "native": {
            "type": "string"
          }
        }
      },
      "MediaResponseDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "ANIME",
              "MANGA"
            ]
          },
          "title": {
            "$ref": "#/components/schemas/MediaTitleDto"
          },
          "synopsis": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "season": {
            "type": "string",
            "nullable": true
          },
          "seasonYear": {
            "type": "number"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "bannerImageUrl": {
            "type": "string"
          },
          "colorHex": {
            "type": "string"
          },
          "episodes": {
            "type": "number"
          },
          "chapters": {
            "type": "number"
          },
          "volumes": {
            "type": "number"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "averageScore": {
            "type": "number"
          }
        },
        "required": [
          "provider",
          "externalId",
          "type",
          "title",
          "status",
          "genres"
        ]
      },
      "MediaListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaResponseDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "SingleMediaResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MediaResponseDto"
          }
        },
        "required": [
          "data"
        ]
      }
    }
  }
}